PHIOS-6962 [Viettel PoC] Timeline RoadScore at 100 is shown as 1

+++ MojioSDK/Models/Vehicles/Trip.swift
@@ -73,8 +73,8 @@ public protocol TripModel: Codable, PrimaryKey {
     var createdOn: Date? { get }
     var lastModified: Date? { get }
     
-    var driverScore: Int? { get }
-    var averageDriverScore: Int? { get }
+    var driverScore: Double? { get }
+    var averageDriverScore: Double? { get }
     
     var properties: P? { get }
 }
